requirements.txt
flask==3.0.3
requests==2.31.0
beautifulsoup4==4.12.3
google-generativeai==0.7.2
python-dotenv==1.0.1
